+++ docs/en/changes/changes.md
@@ -52,6 +52,7 @@
 * Deprecated metrics query v2 protocol.
 * Deprecated record query protocol.
 * Add component ID for go-redis.
+* Add OpenSearch 2.8.0 to test case.
 
 #### UI
 * Revert: cpm5d function. This feature is cancelled from backend.
@@ -82,6 +83,8 @@
 * Fix: the Instance List data display error.
 * Fix: set topN type to Number.
 * Support Metrics Query Expression(MQE) and allows users to do simple query-stage calculation through the expression.
+* Bump up zipkin ui dependency to 2.24.1.
+* Bump up vite to 4.0.5.
 
 #### Documentation